+++ alicloud/resource_alicloud_security_group.go
@@ -55,11 +55,12 @@ func resourceAliyunSecurityGroupCreate(d *schema.ResourceData, meta interface{})
 		return ecsClient.CreateSecurityGroup(buildAliyunSecurityGroupArgs(d, meta))
 	})
 	if err != nil {
-		return err
+		actionName := buildAliyunSecurityGroupArgs(d, meta).GetActionName()
+		return WrapErrorf(err, DefaultErrorMsg, "security_group:", actionName, AlibabaCloudSdkGoERROR)
 	}
 	resp, _ := raw.(*ecs.CreateSecurityGroupResponse)
 	if resp == nil {
-		return fmt.Errorf("Creating security group got a nil response.")
+		return WrapError(fmt.Errorf("Creating security group got a nil response."))
 	}
 	d.SetId(resp.SecurityGroupId)
 	return resourceAliyunSecurityGroupUpdate(d, meta)
@@ -97,7 +98,7 @@ func resourceAliyunSecurityGroupRead(d *schema.ResourceData, meta interface{}) e
 
 	tags, err := ecsService.DescribeTags(d.Id(), TagResourceSecurityGroup)
 	if err != nil && !NotFoundError(err) {
-		return fmt.Errorf("[ERROR] DescribeTags for security group got error: %#v", err)
+		return WrapError(err)
 	}
 	if len(tags) > 0 {
 		d.Set("tags", tagsToMap(tags))
@@ -115,7 +116,7 @@ func resourceAliyunSecurityGroupUpdate(d *schema.ResourceData, meta interface{})
 	args.SecurityGroupId = d.Id()
 
 	if err := setTags(client, TagResourceSecurityGroup, d); err != nil {
-		return fmt.Errorf("Set tags for security group got error: %#v", err)
+		return WrapError(err)
 	} else {
 		d.SetPartial("tags")
 	}
@@ -134,7 +135,7 @@ func resourceAliyunSecurityGroupUpdate(d *schema.ResourceData, meta interface{})
 			return ecsClient.ModifySecurityGroupAttribute(args)
 		})
 		if err != nil {
-			return err
+			return WrapErrorf(err, DefaultErrorMsg, d.Id(), args.GetActionName(), AlibabaCloudSdkGoERROR)
 		}
 		d.SetPartial("name")
 		d.SetPartial("description")
@@ -153,7 +154,7 @@ func resourceAliyunSecurityGroupUpdate(d *schema.ResourceData, meta interface{})
 			return ecsClient.ModifySecurityGroupPolicy(args)
 		})
 		if err != nil {
-			return fmt.Errorf("ModifySecurityGroupPolicy got an error: %#v.", err)
+			return WrapErrorf(err, DefaultErrorMsg, d.Id(), args.GetActionName(), AlibabaCloudSdkGoERROR)
 		}
 		d.SetPartial("inner_access")
 	}